Desafio 6

+++ desafio1.sql
@@ -7,7 +7,7 @@ USE SpotifyClone;
 CREATE TABLE IF NOT EXISTS plano(
   plano_id INT PRIMARY KEY NOT NULL AUTO_INCREMENT,
   plano VARCHAR(100) NOT NULL,
-  valor_plano INT NOT NULL
+  valor_plano DECIMAL(5, 2) NOT NULL
 ) ENGINE=InnoDB;
 
 CREATE TABLE IF NOT EXISTS usuario(
